Created Admin Warning Email Feature.
Implemented an automated cron job feature which identifies data rows with PAYHERESTATUSCODE = 2 and where License key column is empty. That means even if the customer has paid for a kaspersky product, the license key has not generated in KSS Backend.

So this admin-warn email feature will notify admins immediately via email once every half an hour if there's a customer with no license key.

A warning email is only valid for the day which the issue came. So adding new admin emails in future will not gonna receive past admin-warn emails which are already investigated in the past itself.